dev-requirements.txt
altgraph==0.17.4
Babel==2.15.0
contourpy==1.2.1
cycler==0.12.1
fonttools==4.52.4
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tkcalendar==1.6.1
tzdata==2024.1
utils==1.0.2
